[CLEANUP] Remove dividers2tabs functionality
The default for any installation and the whole
core is to use dividers2tabs enabled by default.

Disabling "dividers2tabs" results in a long list
of input fields, which is not usable today
anymore.

The patch removes the option completely
in FormEngine, and removes the option in
ctrl section of the core tables, meaning
that tabs will also be shown in FormEngine.

Releases: master
Resolves: #62833
Change-Id: I2e59fc4a4619a2611859751985da46dd73a9ca0b
Reviewed-on: http://review.typo3.org/33958
Reviewed-by: Stefan Froemken <froemken@gmail.com>
Tested-by: Xavier Perseguers <xavier@typo3.org>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Xavier Perseguers <xavier@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benjamin Mack <benni@typo3.org>