[TASK] Refactor SaltFactory
The patch deprecates static SaltFactory::getSaltingInstance()
and replaces it with the two new non-static methods.
The ->get() method returns a hash instance to check a given password
against a given hash, and ->getDefaultHashInstance() which returns
an instance of the configured default hash method to calculate
a hash for a new password.

The new methods are now strict, non-static and throw exceptions if
something goes wrong. This simplifies mocking in tests and sanitizes
password hash handling in the core.

Change-Id: I186576593202cb6d052bc7c1ca6f81314eddbaf2
Resolves: #85796
Releases: master
Reviewed-on: https://review.typo3.org/57847
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>