[TASK] Cleanup setup module
Some cleanup of the SetupModuleController:

Remove an unused function for adding Javascript,
because it has not been executed since at least 7.6
and there are other ways to add custom Javascript
nowadays. For example with an event, introduced
in #91132, or directly via PageRenderer.

Remove an unused parameter, which has been
forgotten in #93093.

Remove two superfluous typecasts.

Resolves: #93896
Releases: master
Change-Id: If4e4eb5352e46936332a687639399b8edeff0abc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68755
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>