[FEATURE] Add PSR-14 event to enrich password validation ContextData
A new PSR-14 event `EnrichPasswordValidationContextDataEvent` is
introduced to allow extension authors to enrich the `ContextData` DTO
before a password is validated against a password policy.

Additionally, the `ContextData` DTO in the `PasswordReset` class
now includes the backend user username and realName attribute,
which can be used in custom password policy validators.

Resolves: #100294
Releases: main
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: I395a0eb75749c12c4beac3611c436209c07844cf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78259
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>