[TASK] Update display and markup of user settings form
Previously, the form fields in user settings form were displayed in
their entirety, filling the screen.

With this patch, the width of the input mask is now reduced to create
a uniform appearance with forms of the form engine. Some individual
spacing between the form fields has also been adjusted to match those
of the form engine.

The buttons for adding or removing an avatar image now each contain
speaking labels that were previously only aria-labels.

Additionally, the form fields markup in the user settings has been
updated to fully match the FormEngine form field markup.
With this we create the prerequisite that rollouts of global style
adjustments for form fields also work on this form.
Mid-term, this module should be using FormEngine for rendering anyway,
so this is kind of a pre-patch.

Resolves: #100605
Releases: main
Change-Id: Ia83c9d42bcdb2faa12e1c1918d19e66a43f7ad66
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78661
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>