[BUGFIX] Handle unavailability of (dis)allowed extensions in FileBrowser
For cases when the FileBrowser doesn't receive any information
of allowed or disallowed file extensions via the "bparams" query
parameter, the code is slightly adjusted to handle this case
to avoid PHP errors.

Additonally, the avatar field of the user settings is adjusted
to provide corresponding information, while the allowed types
are set to the "imagefile_ext" list. Previously used "allowed"
option in the configuration of the field is removed as it had
no effect. The avatar image is always checked against the
"imagefile_ext" list.

Change-Id: Idfe946a5327e98958cd649ce4be18c729ab4cd9b
Resolves: #101342
Related: #101231
Releases: main, 12.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79987
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>