[TASK] Avoid using sys_log DB field details_nr
"$details_nr The message number. Specific for each $type and
$action. in the future this will make it possible to translate
error messages to other languages"

This never happened.

Field `details_nr` in table `sys_log` is nearly completely
unused: CLI command `syslog:list` only outputs it in verbose
mode, the belog module doesn't at all. There is no information
about the integers being logged, many consumers fall back to
`-1` or `0`, some use a number that looks rather random.

There is a single use in ext:belog related to the "delete similar
errors" functionality:
"If the detailsNo is 11 or 12 we got messages that are heavily
using placeholders. In this case we need to compare both the
message and the actual log data to not remove too many log entries."
11 and 12 are used 4 times. Nowadays most log rows "heavily use"
placeholders, the exception for 11 and 12 doesn't make sense anymore.

The patch removes the entire handling of `details_nr` including
its `sys_log` table column. The API methods are changed and
annotate null as value so PHPStan finds places that still set
some value.

The DB field did default to zero, so it does not harm to not
insert a value during write. This is good for instances that
do not remove the column immediately.

The patch requests to be backported to v13: It should not harm
in existing instances during upgrades and it helps with further
v13 maintenance when main vs. 13.4 are kept identical with this
patch that changes many single places. The v13 version of the
patch will however keep the DB field itself to zero any chance
of a breaking change in case some extension inserts rows manually.

Resolves: #106100
Releases: main, 13.4
Change-Id: Ifef7d8aa2f430817c771d2c71e47575b5556a185
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88069
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>