[BUGFIX] Revert edge case language code ch in setup module
The simplified language "ch" needs a special treatment
in the setup module.

In the future, we should actually remove "ch"
as we know if from TYPO3 (Locales.php) and allow "ch"
to be Chamorro to be in use soon - this would need
an upgrade wizard and some ecosystem changes,
for this reason, we choose this solution for the time
being.

Resolves: #106125
Releases: main, 13.4
Change-Id: Ic98bb9ce68ffcf95749965dcc533bbd0fc69d84b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/88156
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>