[BUGFIX] Allow disabling checkboxes with default in user settings
Previously, it was not possible to disable a
checkbox field in user settings, when it had
a default value, set to true, because the
evaluation treated "falsy" values like 0 as
if no value was set.

This change adjusts the logic to distinguish
between null (unset) and false (explicitly set).

Now, if the value is null and a default is
configured, the default is applied. Otherwise,
the explicit value (including false) is respected.

This ensures checkboxes can be properly disabled
even if they default to true.

Resolves: #106996
Releases: main, 13.4
Change-Id: Iae538196b5a7565327068e6dec72173c6cdb1c51
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89932
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Reviewed-by: Oli Bartsch <bo@cedev.de>
Tested-by: Garvin Hicking <garvin@hick.ing>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Oli Bartsch <bo@cedev.de>